# JPEG decoding much faster!