@Twinside 2016-01-02
I am sooo [guilty](https://github.com/Twinside/Juicy.Voxels/blob/07c9bfad9969c510f78222632f58da16211e9db4/src/Codec/Volume/BitIO.hs#L126). I know that what I wrote is suboptimal, and already fixed part of it in an other project. I will probably reformat your code, and integrate back the other stuff, looks good to me :)

After testing against the test files, sadly the change is bad, tests/jpeg/huge.jpg is wrongly decoded :-/