@Chobbes 2016-01-02
Oh darn. I'll look into it later too. I guess the tests don't check that things were decoded properly?

The image I was using was getting decoded properly. My means of debugging was to put traces printing the result, and state (with only the length of the ByteString) in the unpackInt code, and comparing that with the original implementation. They were identical for what I was testing, but maybe huge.jpg has another special case.

Things most likely to be messed up:
  - BoolReader state
  - Ints that are read across byte boundaries
  - The shift amount could be messed up

It's probably an "off by one" type error somewhere >_<. Apologies.