@Twinside 2016-01-02
I got many files well-decoded with your function, but this one was broken. And yeah the tests just cross convert everything, and there is a visual inspection to be done (I should automate a regression checker or thing like that).